+++ CHANGELOG.md
@@ -11,6 +11,7 @@ the following form: YYYY.0M.0D.
 - Disabled an aspect of the coderunner extension
 - Fixed broken links
 - Disabled AI features globally
+- Created a taskTracker that intially allows users to add, complete and display tasks.
 
 ## [2024.12.30]